Divided statement parser in new files.
This was a huge refactor, and it adds the ability to check if a particular parser
should take `return`, `default`, `in`, `return`, `yield` or `await`, when parsing.
This also divides the tests and adds some more, and puts them in the proper folder.

All parsers now implement `TokenParser`, which has been extended a bit. It would make
sense that this commit reduces performance a bit, since it includes new features.